Refactor: create `archive` folder to move migrations whose dependencies are already deprecated
And that cannot be transpiled again.
