+++ archive/README.md
@@ -1,0 +1,4 @@
+# Archive 
+
+This folder contains the migrations that cannot be transpiled (again) because their dependencies are deprecated.
+These migrations have already been transpiled and their CommonJS code is in the [dist](dist) folder.
