+++ archive/legacy-editor-to-editor/src/slate.d.ts
