+++ archive/legacy-editor-to-editor/src/splishToEdtr/normalize-slate.ts
